chore: bump version to v0.3.0 and fix all clippy warnings
- Update version from 0.2.1 to 0.3.0
- Fix 217 clippy warnings across the codebase:
  - Phase 1: derivable_impls, unwrap_or_default, useless_format, etc.
  - Phase 2: string indexing safety, regex in loop, Iterator::last
  - Phase 3: complex types, too_many_arguments, misc warnings
- All tests passing, zero clippy warnings with -D warnings -A dead_code

+++ crates/aster/src/agents/mcp_client.rs
@@ -1,6 +1,5 @@
 use crate::action_required_manager::ActionRequiredManager;
 use crate::agents::types::SharedProvider;
-use crate::providers::base::Provider;
 use crate::session_context::SESSION_ID_HEADER;
 use rmcp::model::{
     Content, CreateElicitationRequestParam, CreateElicitationResult, ElicitationAction, ErrorCode,
@@ -209,7 +208,7 @@ impl ClientHandler for AsterClient {
 
         // Apply temperature if provided in the request
         if let Some(temperature) = params.temperature {
-            model_config = model_config.with_temperature(Some(temperature as f32));
+            model_config = model_config.with_temperature(Some(temperature));
         }
 
         // Use complete_with_model to apply the custom model config
